Enhancement of the Viewer cf. notes from Homburg
git-svn-id: file:///home/svn/shoola-omero/trunk@4944 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/measurement/view/MeasurementViewerModel.java
@@ -379,7 +379,7 @@ public class MeasurementViewerModel
 		this.magnification = magnification;
 		if( state != MeasurementViewer.NEW)
 			drawingView.setScaleFactor(magnification, new 
-						Dimension((int)getSizeX(), (int)getSizeY()));
+						Dimension(getSizeX(), getSizeY()));
 		else
 			drawingView.setScaleFactor(magnification);
